chat input: fix focus loss on mobile.
Not clear why this isn't always triggered on desktop, but adding `chatInfo.blocks` to the dependency array for the onSubmit useCallback was causing the ChatInput to lose focus after send on mobile after every message. I changed the dependency to `chatInfo.blocks.length` and that seems to have solved it. I also cleaned up the if statements at the top of onSubmit.

Fixes #2506