heap: fix for authors not receiving notifications for comments
Fixes #1872 by making sure that we emit a yarn in this particular instance. They would receive notifications if they were to follow up to the comment, but they weren't getting notified on the initial comment because their patp wasn't in the replies. Now they'll always be notified on any comment on their heap posts.

One question for @arthyn about why it's necessary to return he-core twice for this to compile.