notebook refs: more fixes
This fixes #2526 by preventing refs from being rendered within a notebook reference, also fixes some issues around the way we were calculating remainingChars (it was getting reset early). Also limits the number of images that can be rendered in a ref (to one), handles header text (which users can apparently add by pasting in content from somewhere else, this is not available directly in our UI) and adds an ellipsis to the last inline string in the truncated text.