talk-cli: fixed stack trace related to old wire
We were getting a `%bad-agent-wire` trace because `%talk-cli` wasn't actually dropping the old `/chat/ui` wire during the state transition.

A leave card was included in the subscriptions update, but it wasn't actually leaving because the leave card referred to the aliased `our.bowl` (`+our-self`), without an appended `tc`. Instead of `our.bowl`, it was defaulting to `~zod` each time. It could not find the arm; however, why would it not give a compile error in this case?

Appending `tc` fixed it.