talk-cli: ;chats search refinement, plus fuzzy search
I followed Mark's suggestion to make the search dynamic and use `+find` to check user input against rendered targets, which _significantly_ reduced the line count for the +chats arm. Overall, this is a much more intuitive implementation.

Also, required changing +get-chats to scry for the full $chat map so we can check their corresponding group in +chats. +targets-exists was the only other arm that uses +get-chats - just checking for existstence - so there was minimal breaking changes to fix.