talk-cli: renders pretty chat title next to flag + group pretty title search
The _chat_ pretty title doesn't come in our chat meta data so we scry for the `$chat` to get its corresponding group flag to then scry for the `%channel:group` which contains the pretty title. This logic naturally fit within +meta:tr.

I also added a +match-group arm in +chats to search for a  _group_'s pretty title. Was thinking of passing through the new %flag branch in +meta:tr, but there's no good way to distinguish a chat flag from a group flag and right now we don't render group flags so I figured a helper arm was best added up here.