notify: fix issue with encoded params in post to twilio
We were unnecessarily encoding params before creating a base64 string to use as a signature to post to twilio.
This didn't matter in the past because the device IDs sent by apple devices didn't include any special characters.
Android device IDs include a colon, and the url encoding was replacing the `:` character, which caused the signature created by using `data` to not match the one twilio expected.