talk-cli: popped +order out of +show-targets
In preparation for adding closest-match sorting in +chats and +dms. Non-search prinouts in +dms and +chats still order alphabetically; so does the base case of +view. Best now to have arms passing target lists to +show-targets do their own sorting beforehand.

I also renamed +order to +order-targets for clarity, and changed the target-type ordering to the following:
- clubs go first
- then flags
- then ships
So that ships are printed closest to the prompt. Felt like this was more user-friendly, and is closer to the pattern we'll follow with closest-match sort which will be prompt-centric.