diary: fix issue with clearing all note text
The useEffect we were using to load initial text into the note was overriding the user's ability to clear out all of the note text. This PR adds some local state to prevent that from happening. Also switches to using tiptap's isEmpty property to check if the notebook is empty, rather than checking for an empty string, which is what tiptap's docs suggest.