notebook: block content menu usability
fixes LAND-598 by updating the menu style to match the rest of our app, adding a "+" button to access it, updating the placeholder text, ensuring that the placeholder text appears on each new node.

Also updates tiptap to the latest stable version (2.0 is out of beta). Required pinning some specific prosemirror deps to get typescript to be happy.