chat: display status indicator for pals
If a ship is known in the pals state, we display a status indicator
similar to the one in the original pals interface. Tooltip displays the
tags/lists for that pal, if outgoing.

Since we already have the pals data (for use in leap), this is simply
the addition of an indicator component.