profiles: hide pals action if not installed
The conditional for this was implemented incorrectly. Additionally, the
styling would not account for it.

Alongside the fix, we refactor the `!contacts` branch into a single main
rendering branch, within which we check that condition to decide whether
to include a component. Doing so prevents us from duplicating the now
slightly more complex button rendering logic, and will make additions
and modifications easier going forward.

Fixes LAND-643.