WIP: handle case where notebook host is offline
This allows a user to create a post while the notebook host is offline or unreachable.

The post gets stored in the cache and we add the temporary noteId to an array of pendingNotes.

We check if a noteId is in that array, if it is the note headline in the notebook list his grayed out and a warning is shown on the note itself when viewed.

We use vitals to check if the host is online. If they are, and if we have pendingNotes, then we fetch the notes from the live server and compare it to the length in the cache. If the length matches, we guess that the server now has the note, so we do a refetch from the server to replace the cache and clear the pending notes.

Need a more robust way to make sure to only refetch if the actual note is available on the server, otherwise the user will lose data on the refetch.

fixes LAND-673