lure: autojoin from any view
Before this change, the auto join logic was only executed on the Home Route. Unfortunately, we immediately drop new users into the Find Groups view, so this effect was not being executed.

This PR extracts the logic to a dummy component that's nested under the root router. Thus this will now work for wherever a newly onboarded user first opens Groups.

This along with https://github.com/tloncorp/ylem/pull/1597 resolves GRO-598