wip: partially update diary for backlog windowing
Here we update /app/diary's update processing logic to match the recent
changes made to the diff type, and account for the
not-yet-fully-implemented backlog windowing system.

We also add a /lib/sparse containing types & utilities for sparse atom
sets. In the future, this may be made fully generic, but for now, it
fits our use case just fine.

As always, uncompiled & untested, be forewarned.