diary: make load logic compile
The new diary version has a different $net type, which excludes the
publisher case. This makes sense for /app/diary, but not necessarily for
/app/diary-server. Luckily, the server never refers to this field!

For diaries which we were publishing, we make up a $net with a %chi
$saga. The /app/diary's +inflate-io will still want to open a
subscription to the diary (even where the host is the local ship).
Specifying a %chi $saga lets it do so unimpeded. Being local lets us
guarantee that this Is Fine™.