negotiate: add tests, fix discovered bugs
We add automated tests for the version negotiation library. These are
written in the style of the Eyre tests, where we keep passing the test
agent forward as we do operations on it, extract cards and other state.

The tests revealed a couple of issues, primarily around order of
operations. Here we include fixes, so that all tests pass.