negotiate: correctly handle kicks & watch-nacks
For subscriptions made by the inner agent, we already track %watch and
%leave cards as "intent to establish this subscription" in the .want
map. However, receiving kicks and watch-nacks means gall has rescinded
the agent's intent to subscribe on its behalf, and we should handle them
as such, removing the relevant subscription from the .want map, only
re-adding it if the inner agent explicitly resubscribes.

Because +on-agent only knows the wire, and not the full gill of the
relevant subscription, we are forced to add a library prefix to the wire
of subscriptions established by the inner agent. We "pack" and "trim"
the wire before emitting cards and injecting events respectively, and
read the gill out of the wire whenever a subscription gets kicked or
nacked.