negotiate: better nack handling of version subs
Previously, if a version subscription got nacked, we would still
maintain the last-heard version in state. Since the subscription failed
to establish, we no longer actually know, and so risk keeping an old
version indication in state.

Now, when receiving a watch-nack on a version subscription, we consider
the version for the relevant gill+protocol to be "unknown", updating the
state as such, and processing any mismatch side-effects.

Additionally, if a version subscription nacks, and we no longer have a
need to know, we simply don't reestablish the subscription. We
intentionally don't do this for kicks. A watch-nack is a much stronger
signal, and if we can keep the subscription in place, we should: it
might become relevant again later.