negotiate: don't track & re-wrap lib-managed subs
The library should not track library-generated subs in the .want map.
Such subs have extended wires and should already be tracked by the
library.

The library should also not re-wrap/pack wires of subs it aims to kill,
because the wires we're killing on come straight out of the bowl, and
might already be the packed ones.