lib: test-agent, testing harness for agents
/lib/test-agent is a library that helps you write tests for agents in
the monadic ;< style. an agent and its bowl are continuously "passed
forward" as you perform operations on them. the library partially
manages the bowl, making sure that subscriptions are reflected
accurately based on cards coming out of, and events going into, the
agent being tested.

comes with helpers for running agent functions, testing emitted cards
for equivalence, retrieving data from the agent, and so on. for an
example, see the header docs of the library file.

most of the logic herein cribbed from the tests for /lib/negotiate,
which will be updated to use this library in a following commit.