negotiate: refactor tests with /lib/test-agent
Along the way, we discover a bug where we were killing version
negotiation subscriptions when we started mismatching with the agent we
were negotiating with. We fix that too.

(Additional bugs were found during the refactor, but since they were
partially caused by poorly written tests, we delegated that to an
earlier commit, 542a87c7.)