negotiate: only block pokes if we know we mismatch
Eagerly crashing on pokes to mismatching agents is essentially an
optimization. If we don't actually know their version(s) yet, we may
match, and so sending a poke may be safe. Blocking the poke in that case
is an unnecessarily eager hindrance.

So, here we let pokes through unless we know for sure that we mismatch.
If we aren't sure, this suggests we don't know a version for them yet,
so we should initiate version negotiation if we haven't yet.