channels: correct %leave after backlog/checkpoint
We are supposed to issue a %leave for backlog & checkpoint
subscriptions, so that the subsequent %kick sent by the server is
dropped instead of triggering our resubscribe logic.

But we were sending a %leave to ourselves, not the server agent. This
changes that to send the %leave to the agent we were subscribing to in
the first place.

Fixes LAND-1150.