chat: use different message ids during migration
Chat had two notions of a message identifier: one that comes packed in
the message itself, of the shape `[@p @da]`, and one that is used as the
index for the ordered map of messages, of the shape `@da`. Channels, on
the other hand, only ever used the same `@da` for both cases.

During migration, chat was assigning messages their mop-index as their
identifier.
Chat used to encode references to messages in a path containing the
`[@p @da]` identifier.
The `@da`s from the two kinds of identifiers were generally not exactly
the same.

The combination of those three factors meant that post-migration,
references would be pointing to non-existent messages, and fail to
resolve. (We were also not updating the reference path syntax to the
newer version, but that alone is trivial.)

Here, we make chat pass on the `@da` from the `[@p @da]` identifier
during migration. These are technically not guaranteed to be unique, but
in practice should be unique 99.999…% of the time.
This then lets us easily update the message reference paths into the new
format, after which they will continue to resolve like they used to.

(The alternative was to somehow update the message reference paths to
contain the mop-index id instead, which we might not have been able to
retrieve 100% of the time.)

This does not fix the "bad" migration performed on ships that are
already running this branch. Their pre-migration references are busted,
but that's relatively benign. Re-running the migration fixes this, at
the expense of wiping post-migration content.

Resolves LAND-1135.