negotiate: ensure *all* subs are wrapped
Previously, we weren't forcing the inner-agent to re-establish its
subscriptions upon getting wrapped in the library, instead letting it
leave subs without version requirements in place. However, this means
that there is not enough information in the wires for the library to do
proper bookkeeping about what subscriptions the inner agent wants open,
upon receiving %kick or negative %watch-ack signs.

Here, we update the "initial on-load" logic to simulate kicks for *all*
the inner-agent's existing subscriptions. We capture and wrap the
subsequent %watches (if any) as normal.

Includes shoddy migration logic for agents running earlier versions of
the library. It assumes no versions are changed during the upgrade,
which we know to be true because we aren't making any changes here, and
know the small set of ships running with the old version of the library.
It's ugly though, and we should excise it sometime soon.