chat: revert eea49cd, update refs in-place instead
eea49cd had taken the too-bigbrained approach to helping refs cross the
migration by updating the message identifiers to match the old ones.

Here, we reject that approach and all the complications it brings, and
instead do a best-effort in-place updating of references instead. For
every chat message reference, we update it to point to the new chat
message id. This includes, for threaded messages, seeking out the id of
the parent message and including that in the reference path.