heap, diary: use blank wire for maintenance pokes
In their +on-load logic, they would emit pokes for re-checking perms and
leaving unnecessary subscriptions. These were done with a "bunted"
channel wire though, resulting in crashes when trying to handle the
eventual poke-n/ack.

We replace the wire with /, which gets handled as a no-op in all cases.
This way, we don't have crashing output during migration (or other
loads).