negotiate: simulate kicks for killed subscriptions
When we kill subscriptions due to version mismatch reasons, we would
previously only emit a %leave, and silently re-establish the
subscription once versions matched again.

This was problematic behavior for agents that want to subscribe to "last
we heard and onward", which is a common pattern in chat-style
applications, among others. They want to be notified about the
subscription being dropped, because they want to re-establish the
subscription on a different path, to avoid receiving duplicate updates.

Here, we update the library to, when proactively %leave-ing
subscriptions, also remove them from the list of desired subscriptions,
and inject a %kick into the underlying agent. Whatever subscriptions the
agent tries to establish while we're in a mismatching (or not yet known)
state we do still hold on to as normal, and release as %watch cards once
versions start to match again.

Updates tests to match, includes some extra state checks to make sure
this all operates as expected.
Tests diff is a little bit noisy, since we now want/need to dinstiguish
between subscribes and *re*subscribes.