channels, chat: include threads in unread state
Unread bookkeeping was not accounting for new messages added into
threads/as replies. This left the client unable to render accurate
un/read markers for threads and their contents.

Here, we book-keep whenever new (and thus, unread) replies appear.
Whenever we recalculate the full "unread state" for the client, we use
the marked threads to include them in the unread count & pointers.

This upgrades cleanly because we put the new (set) in place of a
vestigial ~ (in the $remark type).

This might not operate correctly in the presence of deleted messages,
but previous behavior was broken in the same way. See also the added
TODO.

Both channels and chat are changed here. The changes compile, but the
frontend has not been updated to match, and as such these backend
changes have not yet been tested thoroughly.