lib: don't bundle dbug & verb in negotiate
Developers should have the flexibility to determine whether they want
verb and dbug "above" or "below" the library.

For the groups agents that use it, we put it "below" for now, providing
a view of what the agent is being told is happening, rather than the
orchestration the library is providing.