channels: revision reply contents, support editing
Channels already supported editing posts. This would increment the
revision number stored for their content in state. For replies, no such
affordance or revision number existed.

Here, we add an %edit case to the available reply commands, and add a
revision number into the state for replies to support this.

Because this is a protocol change, we bump the protocol version number.
Version negotiation will automagically do its thing.