groups: expose "can-read" as gate through scry
Instead of only exposing a per-channel per-ship scry endpoint for
checking read permissions, we now also expose a generic endpoint that
produces a gate. This gate can be used to run a read permissions check
with varying arguments repeatedly.

We update the +can-read:perms helper from /lib/channel-utils to use this
gate, and make sure its scry of the gate is memoized. This removes the
scry overhead for subsequent invocations within the same event.

There's light concerns here around the context of the gate being used to
exfilterate otherwise-private data from the agent. However, two reasons
not to care:
- These endpoints are only accessible locally.
- The groups agent already exposes nearly all of its state on other scry
  endpoints, and userspace permissions don't exist yet, so this won't be
  "leaking" data that isn't already accessible.