various: remove old/stale/unused dependencies
We remove all references to and dependencies on graph-store,
metadata-store, group-store, and versions of channel types older
than the previous generation.

This does mean that we remove migration/import logic for versions of
groups preceding the previous one, but the migration path coming from
that far back was already known to need some manual guidance.

Notably inclusion is a bugfix in /app/notify: it was still scrying into
%group-store, when it should have been updated to scry into %groups
instead.