profile: clear cache on unbind, give up on /
We make sure to clear the cache whenever we unbind, to prevent the
profile from unintentionally sticking around.

Unfortunately, this means the /profile url will just serve a static 404,
instead of falling back to whoever had bound the catch-all at /. This is
probably an eyre & runtime misfeature, and we make note of it as such.

We also remove code relating to binding the profile at /. /profile is
plenty fine for the time being, and with the above, binding at / would
actually be actively dangerous to an urbit.