scroller: invert list based on reading direction
Fixes LAND-1353.

This changes our logic about whether or not to invert the list. Rather than relying on load direction, we make the invert/uninvert decision base on reading direction (whether the user perceives to be scrolling up or down). If scrolling up, we invert. If scrolling down, we "uninvert". This avoids jumpiness associated with scrolling "up" in an uninverted state, or "down" in an inverted state.

This also moves a few variables around for easier readability, and moves a few variables into useMemo where it seems appropriate.