s3: allow video uploads
This PR fixes LAND-1480 by allowing users to upload videos.

We already supported rendering video content in chat and galleries, we just didn't support user uploads for it.

This PR simply adds similar functionality to what we already had for image uploads, for video.

There is on hacky part, you'll see a TODO about it in ChatContent.tsx.