unreads: use chat store unreads for sidebar count
Fixes LAND-1574.

We were using the unread state synced from the backend for the unread count on the messages tab button.
This caused a delay in updating the count due to the delay in the updated state.

If we switch to using the chat store unreads, the count updates immediately.