Search: Don't focus ChatInput for /dm/groups/... routes
Since the nav bars of groups and talk were combined, when viewing a Group Chat using the Messages tab on desktop, the ChatChannel view does't recognize the search route if  the Search button is clicked.
In this case, the URL has a "/dm/groups" portion (instead of "/groups" as when in the Groups nav tab).
Chat input should not be focused in those cases to allow the search input to be used.