chat: make sure data is up to date when we have unreads
Fixes LAND-1556.

getQueryData was always returning undefined in my testing locally, so rather than only refetching if we have cached data for a query, we just fetch the query if we have an unread for the channel (since we know we *will* use that data later).

I also noticed that channelKey in state/channel/keys.ts was using 'channel' rather than 'channels'. 'channel' isn't used as a queryKey anywhere afaict, I assumed this was supposed to be 'channels', so I changed that.