channels: support non-knot text in searches
On the backend, we were taking the search query directly out of the
path. Assuming all paths are well-formed, this meant we had no way of
receiving text that didn't fit into the path's @ta (aka knot) character
limitations.
Here, we first try to decode the query part of the search scry path as a
knot-encoded string. Only if that fails do we take the path element
as-is. (That way, we maintain backwards compatibility, and continue
supporting trivial search cases trivially.)

On the frontend, we were injecting the search query into the scry url
directly, even if it contained non-url-safe characters. Something
further down the stack would url-encode the url string, but the backend
wasn't prepared to handle this (and shouldn't have been).
Here, we add a "stringToTa()" utility for knot-encoding strings, and
make sure to do that to the query before putting it into the url.

With the combination of these, we now support arbitrary text (including
unicode!) in the search.