channels: new search api, bounded by msgs searched
The existing search api is bounded by the amount of results returned.
This makes searching for something with few results a pathological case:
we will continue searching _all_ messages in a channel until we've hit
the desired amount of results, or exhausted the entire backlog. This
results in search being slow in unpredictable ways.

We add a new search api (that is, a new scry endpoint) where the search
is bound (and paginated) by the total amount of messages searched,
rather than the subset that gives results. This way, we can bound the
amount of computation that happens for any given "page" of search,
making for more predictable ux. It also explicates the distance to which
the backlog has been searched (by including a "last-searched" message
id), which allows for indicating search progress even in the absence of
search results.

To maintain a simple api, we bound by the number of top-level messages,
rather than messages plus replies. In contexts where replies are
expected to make up the bulk of content, the search bounds/window should
be adjusted downwardly to compensate.

Frontend search code has been updated to use this new endpoint, but this
has revealed inconsistencies in its ability to continue searching,
detecting when it needs to, etc. Further work there is needed.

(Note also that this doesn't yet apply similar changes to /app/chat.
Those will follow once things have settled for channels.)