vite: preserve empty path segments when proxying
The http proxy used by vite eats empty path segments.
(See also http-party/node-http-proxy#1420.)
However, empty path segments are legitimate according to the URI spec
(RFC 3986) and not unheard of in urbit APIs.

To work around the problem, we "hide" empty path segments from the proxy
by rewriting the paths, and then make the proxy re-rewrite the paths
before emitting the request.

Of course, if we ever want to make legitimate requests to paths
containing @@@, we'll be in trouble, but that seems unlikely enough.

(Arguably the pattern for the proxy added here should capture _all_
requests, rather than just the ones containing empty path segments, so
as to make all requests follow the same codepath. Unclear at this time
whether that actually matters or not.)