tests: refine scry mockery
The mock scry gate was producing raw nouns, which meant the best we
could do was let type-based scry failures remain opaque.

Now, we enforce that vases are produced, and we leverage the full input
to the scry gate, to verbosely warn about scry result type mismatches
when they occur.

The hacks around ;; being slow are... somewhat suspect, but in practice,
in these particular cases, they should always be safe.