perf: speed up sidebar sorter
I was getting annoyed with how slow the sidebar can be, so I did a performance profile and root caused that this particular function was pretty heavy.

Saw a marked improvement in performance for this function by just pre-computing the values for comparison before running the comparison and eliminating the use of .reverse().

Change should be most noticeable when navigating back to the main groups list from a group, or just loading the app initially. There's still some slowness with the group sidebar itself, which we should root cause at some point.