e2e: remove specs interdependency
005-mobile-chat-options.spec.ts expects that a chat message from spec 001 is present.
This can cause random failures if spec 001 has not finished before 005 starts.
Also, specs should be independent so that they can be run individualy.

This change consists in adding one step to spec 005 where the chat message is created. The test then continues from the 2nd ship by reacting to that message.