chat: fix missing reply meta on posts in outline requests
We were attempting to re-use the `uv-post` arm for `uv-post-without-replies`,
but in order to do that we were setting replies to null before calling `uv-post`.
This meant that the get-reply-meta arm, when being called from `uv-post`,
would show no reply-meta for the post, even if there were replies.

Fixes LAND-1673