groups: don't crash during "can read" on-peek
The scry endpoints for a group's channels' permission checks would crash
if the channel did not exist in that group. In cases where the channel
was deleted, other agents might still know about the channel, and so
still try to ask the group about relevant permissions, but the scry
handling would crash.

Here we update the scry logic to not crash, and instead produce a "not
allowed to read" flag in cases where the asked-about channel doesn't
exist in the group.

We suspect this will ameliorate at least some of the problems seen
around role modifications and permission rechecks, as this crashing scry
was causing group fact handling to crash, causing locally-originating
kicks, causing resubscribes.