channels: /latest & /[channel]/changes endpoints
/latest gives you the recency & latest message for every channel.
Optionally, provide a timestamp, to only receive results that changes
after that time.

/[channel]/posts/changes gives n amount of messages that changed before
a given timestamp, moving backwards from there. This is subtly different
from the existing /older endpoint, in that it also accounts for edits,
reactions, and other non-creation changes to messages.

Untested and still somewhat a work in progress. There's unanswered
questions and un-marked types here, at least.