channels: proper output marks for new endpoints
For the /[channel]/posts/changes, we simply fall back on %channel-posts,
where the .older value indicates the oldest time included, if we didn't
reach the furthest back we could.

For /heads (formerly /latest), we add a new %channel-heads mark, a list
of per-nest recency and latest message, if any. We implement appropriate
json conversions.